// src/lib/api/v1.d.ts
export interface components {
    schemas: {
// ...
        TransitRoute: {
            /**
             * @description Entitás azonosító
             * @example BKK_DF544
             */
            agencyId: string;
            /**
             * @description Kerékpár szállítása engedélyezett-e a járaton.
             * @example false
             */
            bikesAllowed: boolean;
            /**
             * @deprecated
             * @description A járat színe. Deprecated: használjuk a `style` attribútumot helyette.
             * @example FFD800
             */
            color: string;
            /**
             * @description A járat neve. Ha egy | szerepel a névben, akkor a végállomásokat választja el amelyek külön sorokban megjeleníthetőek. COMPACT referenciák kérése esetében nincs kitöltve.
             * @example Újbuda-központ M | Széll Kálmán tér M
             */
            description?: string;
            /**
             * @deprecated
             * @description A járat ikonjának megjelenítési szövege. Deprecated: használjuk a `style` attribútumot helyette.
             * @example 4
             */
            iconDisplayText?: string;
            /**
             * @deprecated
             * @description A járat ikonjának megjelenítési típusa. Deprecated: használjuk a `style` attribútumot helyette.
             * @example BOX
             */
            iconDisplayType?: string;
            /**
             * @description Entitás azonosító
             * @example BKK_DF544
             */
            id: string;
            /** @description A járat hosszú neve. */
            longName?: string;
            /**
             * @description A járat rövid neve.
             * @example 4
             */
            shortName?: string;
            /**
             * Format: int32
             * @description A járat rendezési sorszáma a többi járathoz viszonyítva.
             * @example 10
             */
            sortOrder: number;
            style: components["schemas"]["TransitRouteStyle"];
            /**
             * @deprecated
             * @description A járat szövegének színe. Deprecated: használjuk a `style` attribútumot helyette.
             * @example 000000
             */
            textColor: string;
            type: components["schemas"]["TraverseMode"];
            /** @description A járat linkje. */
            url?: string;
        };
// ...
        TraverseMode: "BICYCLE" | "CAR" | "WALK" | "BUS" | "CABLE_CAR" | "COACH" | "FERRY" | "FUNICULAR" | "GONDOLA" | "RAIL" | "SUBURBAN_RAILWAY" | "SUBWAY" | "TRAM" | "TROLLEYBUS" | "LOCAL" | "TRANSIT";
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}